pyproject.toml
[project]
name = "static-embedding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "duckdb>=1.2.2",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "polars>=1.29.0",
    "pyarrow>=20.0.0",
    "sentence-transformers>=3.3.1",
    "umap-learn>=0.5.7",
]
